Better Junc
better junction handling. You can now do for the PathOld \??\W:\ and have the replacement path of C and it now replaces it properly. As junctions are always absolute paths and as of the MKLINK command on windows 10 it no longer supports the prefix of \??\ you can correct specific prefixed paths only of \??\ or some other meta old format